requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.1
django-environ==0.12.0
djangorestframework==3.16.0
et-xmlfile==2.0.0
flake8==7.2.0
kombu==5.5.3
mccabe==0.7.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
